requirements.txt
# Django Core
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0

# Database
psycopg2-binary>=2.9,<3.0

# Async Tasks
celery>=5.3,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.5,<3.0
django-celery-results>=2.5,<3.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0

# Utils
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
xmltodict>=0.13,<1.0
Pillow>=10.0,<11.0

# Development
ipython>=8.0,<9.0

# Production
gunicorn>=21.0,<22.0
whitenoise>=6.6,<7.0
